requirements.txt
dgl
pandas
pyvis
Pillow